+++ src/engine/platform/multipcm.h
@@ -92,7 +92,7 @@ class DivPlatformYMF278: public DivDispatch {
     void forceIns();
     void notifyInsChange(int ins);
     void notifyInsDeletion(void* ins);
-    const InsMapping& getInsMapping(size_t ins);
+    const InsMapping& getInsMapping(int ins);
     int calcFreq(int basePitch);
     const char* getEffectName(unsigned char effect);
     bool isStereo();
